+++ lib/svg_sprite_ex/inline_asset.ex
@@ -1,4 +1,4 @@
-defmodule SpriteEx.InlineAsset do
+defmodule SvgSpriteEx.InlineAsset do
   @moduledoc false
 
   @enforce_keys [:attributes, :inner_content]