+++ lib/svg_sprite_ex/sprite_ref.ex
@@ -1,4 +1,4 @@
-defmodule SpriteEx.SpriteRef do
+defmodule SvgSpriteEx.SpriteRef do
   @moduledoc """
   Compile-time sprite-backed SVG reference.