+++ lib/svg_sprite_ex/svg.ex
@@ -1,4 +1,4 @@
-defmodule SpriteEx.Svg do
+defmodule SvgSpriteEx.Svg do
   @moduledoc """
   SVG rendering component.
 
@@ -14,19 +14,19 @@ defmodule SpriteEx.Svg do
 
   use Phoenix.Component
 
-  alias SpriteEx.InlineAsset
-  alias SpriteEx.InlineRef
-  alias SpriteEx.SpriteRef
+  alias SvgSpriteEx.InlineAsset
+  alias SvgSpriteEx.InlineRef
+  alias SvgSpriteEx.SpriteRef
 
   attr :ref, :any, default: nil
   attr :rest, :global
 
   @doc """
-  Renders an SVG from a SpriteEx ref.
+  Renders an SVG from a SvgSpriteEx ref.
 
-  Pass a `SpriteEx.SpriteRef` from `sprite_ref/1` or `sprite_ref/2` to render
+  Pass a `SvgSpriteEx.SpriteRef` from `sprite_ref/1` or `sprite_ref/2` to render
   a `<use>` tag backed by a generated sprite sheet. Pass a
-  `SpriteEx.InlineRef` from `inline_ref/1` to inline the compiled SVG markup.
+  `SvgSpriteEx.InlineRef` from `inline_ref/1` to inline the compiled SVG markup.
 
   ## Examples